+++ Sources/Sentry/SentryUser.m
@@ -47,7 +47,7 @@ NS_ASSUME_NONNULL_BEGIN
     return serializedData;
 }
 
-- (BOOL)isEqual:(id)other
+- (BOOL)isEqual:(id _Nullable)other
 {
     if (other == self)
         return YES;