+++ Tests/SentryTests/Protocol/SentrySdkInfo+Equality.m
@@ -2,7 +2,7 @@
 
 @implementation SentrySdkInfo (Equality)
 
-- (BOOL)isEqual:(id)object
+- (BOOL)isEqual:(id _Nullable)object
 {
     if (object == self)
         return YES;